task3(2)

+++ test.py
@@ -1,6 +1,6 @@
 import numpy as np
 L = np.zeros((4, 5))
-N = np.ones((3, 3))
+N = np.ones((3, 3), dtype=int)
 M = np.random.randint(10, 50, (2, 6))
 print(L)
 print(N)
